Update define domain metadata read method for both define versions and add tests
