+++ cdisc_rules_engine/services/define_xml/base_define_xml_reader.py
@@ -294,19 +294,6 @@ class BaseDefineXMLReader(ABC):
     def _get_variable_is_collected(self, itemdef):
         pass
 
-    def _get_metadata_representation(self, metadata) -> dict:
-        """
-        Returns metadata as dictionary.
-        """
-        return {
-            "define_dataset_name": metadata.Domain,
-            "define_dataset_label": str(metadata.Description.TranslatedText[0]),
-            "define_dataset_location": getattr(metadata.leaf, "href", None),
-            "define_dataset_class": str(metadata.Class.Name),
-            "define_dataset_structure": str(metadata.Structure),
-            "define_dataset_is_non_standard": str(metadata.IsNonStandard or ""),
-        }
-
     def validate_schema(self) -> bool:
         """
         Validates Define XML Schema.
