+++ cdisc_rules_engine/services/define_xml/define_xml_reader_2_0.py
@@ -22,3 +22,17 @@ class DefineXMLReader20(BaseDefineXMLReader):
 
     def _get_variable_is_collected(self, itemdef):
         return self._get_origin_type(itemdef) == "CRF" if itemdef.Origin else None
+
+    def _get_metadata_representation(self, metadata) -> dict:
+        """
+        Returns metadata as dictionary.
+        """
+        return {
+            "define_dataset_name": metadata.Domain,
+            "define_dataset_label": str(metadata.Description.TranslatedText[0]),
+            "define_dataset_location": getattr(metadata.leaf, "href", None),
+            "define_dataset_class": metadata.Class,
+            "define_dataset_structure": str(metadata.Structure),
+            # v2.0 does not support is_non_standard. Default to blank
+            "define_dataset_is_non_standard": "",
+        }
