+++ cdisc_rules_engine/services/define_xml/define_xml_reader_2_1.py
@@ -22,3 +22,16 @@ class DefineXMLReader21(BaseDefineXMLReader):
 
     def _get_variable_is_collected(self, itemdef):
         return self._get_origin_type(itemdef) == "Collected" if itemdef.Origin else None
+
+    def _get_metadata_representation(self, metadata) -> dict:
+        """
+        Returns metadata as dictionary.
+        """
+        return {
+            "define_dataset_name": metadata.Domain,
+            "define_dataset_label": str(metadata.Description.TranslatedText[0]),
+            "define_dataset_location": getattr(metadata.leaf, "href", None),
+            "define_dataset_class": str(metadata.Class.Name),
+            "define_dataset_structure": str(metadata.Structure),
+            "define_dataset_is_non_standard": str(metadata.IsNonStandard or ""),
+        }
